pyproject.toml
[build-system]
requires = ["setuptools>=61,<77"]
build-backend = "setuptools.build_meta"

[project]
name = "novelai-openclaw-adaptor"
version = "0.1.3"
description = "NovelAI shim and image CLI for OpenClaw-friendly usage."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Open Source Contributors" }
]
keywords = ["novelai", "openclaw", "openai", "shim", "cli"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]
dependencies = [
  "novelai-sdk",
]

[project.scripts]
novelai-shim = "novelai_openclaw_adaptor.shim:main"
novelai-image = "novelai_openclaw_adaptor.image_cli:main"
novelai-config = "novelai_openclaw_adaptor.config_cli:main"

[tool.setuptools]
packages = ["novelai_openclaw_adaptor"]
license-files = []
